fix the get cart item to return the needed data

+++ Catalogs.Application/DTOs/ProductDTO.cs
@@ -8,7 +8,7 @@ public class ProductDTO
     public Guid Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public decimal Price { get; set; }
+    public double Price { get; set; }
     public Guid CategoryId { get; set; }
     public string SKU { get; set; }
     public int StockQuantity { get; set; }
